+++ web/CHANGES.md
@@ -1,7 +1,7 @@
 # unreleased 0.3.0
 ## Add
 - enable Rust nightly feature `error_generic_member_access` when `xitca-web`'s `nightly` crate feature is enabled. this enables runtime context type interaction like `std::backtrace::Backtrace` for enhanced error handling.
-- `ErrorStatus` error type would try to capture `Backtrace` when `nightly` crate feature is enabled.
+- `ErrorStatus` error type would try to capture `Backtrace`.
 - `Redirect` and `Html` handler types can be used as standalone route service. Example:
     ```rust
     App::new()